proof/{proof,reachability}: move proof reading methods to leaf class
